Sync meta after default methods
